requirements.txt
Flask==2.3.2
gunicorn==21.2.0
Werkzeug==2.3.6
requests==2.31.0
python-dotenv==1.0.0
jsonschema==4.17.3
Pillow==10.0.0
pydantic==2.0.2
ffmpeg-python==0.2.0
openai-whisper==20231117
Flask-RESTful==0.3.10
